+++ spring/demo/src/main/java/com/example/demo/apply/enums/ApplyStatus.java
@@ -1,0 +1,9 @@
+package com.example.demo.apply.enums;
+
+public enum ApplyStatus {
+    APPLIED,        // 응모 완료
+    SELECTED,       // 당첨
+    NOT_SELECTED,   // 탈락
+    PAID,           // 결제 완료
+    CANCELED,        // 당첨 취소
+}
