--- spring/demo/src/main/java/com/example/demo/base/code/BaseCode.java
@@ -1,6 +1,0 @@
-package com.example.demo.base.code;
-
-public interface BaseCode {
-    ReasonDTO getReason();
-    ReasonDTO getReasonHttpStatus();
-}
