--- spring/demo/src/main/java/com/example/demo/base/code/BaseErrorCode.java
@@ -1,6 +1,0 @@
-package com.example.demo.base.code;
-
-public interface BaseErrorCode {
-    ErrorReasonDTO getReason();
-    ErrorReasonDTO getReasonHttpStatus();
-}
