--- spring/demo/src/main/java/com/example/demo/base/code/ReasonDTO.java
@@ -1,14 +1,0 @@
-package com.example.demo.base.code;
-
-import lombok.Builder;
-import lombok.Getter;
-import org.springframework.http.HttpStatus;
-
-@Getter
-@Builder
-public class ReasonDTO {
-    private final boolean isSuccess;
-    private final HttpStatus httpStatus;
-    private final String code;
-    private final String message;
-}
