--- spring/demo/src/main/java/com/example/demo/base/code/exception/CustomException.java
@@ -1,23 +1,0 @@
-package com.example.demo.base.code.exception;
-
-import com.example.demo.base.ApiResponse;
-import com.example.demo.base.status.ErrorStatus;
-import org.springframework.http.ResponseEntity;
-
-public class CustomException extends RuntimeException {
-
-    private final ErrorStatus errorStatus;
-
-    public CustomException(ErrorStatus errorStatus) {
-        super(errorStatus.getMessage());
-        this.errorStatus = errorStatus;
-    }
-    public ErrorStatus getErrorStatus() {
-        return errorStatus;
-    }
-    public static <T> ResponseEntity<ApiResponse<T>> createErrorResponse(ErrorStatus errorStatus, T data) {
-        return ResponseEntity
-                .status(errorStatus.getHttpStatus())
-                .body(ApiResponse.onFailure(errorStatus, data));
-    }
-}
