--- spring/demo/src/main/java/com/example/demo/entity/Apply.java
@@ -1,34 +1,0 @@
-package com.example.demo.entity;
-
-import com.example.demo.entity.enums.ApplyStatus;
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Getter
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class Apply extends BaseEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "apply_id")
-    private Long id;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "user_id")
-    private User user;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "session_id")
-    private Session session;
-
-    @Enumerated (EnumType.STRING)
-    @Column(columnDefinition = "VARCHAR(15)")
-    private ApplyStatus applyStatus;
-
-}
