--- spring/demo/src/main/java/com/example/demo/entity/enums/EventStatus.java
@@ -1,10 +1,0 @@
-package com.example.demo.entity.enums;
-
-public enum EventStatus {
-    APPLY_NOT_OPENED,       // 응모 전
-    APPLY_OPEN,             // 응모 중
-    APPLY_CLOSED,           // 응모 마감
-    TICKETED,               // 예매 완료
-    IN_PROGRESS,            // 공연 중
-    ENDED,                  // 공연 종료
-}
