+++ spring/demo/src/main/java/com/example/demo/event/enums/AgeLimit.java
@@ -1,0 +1,8 @@
+package com.example.demo.event.enums;
+
+public enum AgeLimit {
+    ALL,        // 전체 관람가
+    AGE_12,     // 12세 관람가
+    AGE_15,     // 15세 관람가
+    AGE_18,     // 18세 관람가
+}
