+++ spring/demo/src/main/java/com/example/demo/global/response/code/BaseCode.java
@@ -1,0 +1,6 @@
+package com.example.demo.global.response.code;
+
+public interface BaseCode {
+    ReasonDTO getReason();
+    ReasonDTO getReasonHttpStatus();
+}
