+++ spring/demo/src/main/java/com/example/demo/global/response/code/BaseErrorCode.java
@@ -1,0 +1,6 @@
+package com.example.demo.global.response.code;
+
+public interface BaseErrorCode {
+    ErrorReasonDTO getReason();
+    ErrorReasonDTO getReasonHttpStatus();
+}
