+++ spring/demo/src/main/java/com/example/demo/global/response/code/ErrorReasonDTO.java
@@ -1,0 +1,14 @@
+package com.example.demo.global.response.code;
+
+import lombok.Builder;
+import lombok.Getter;
+import org.springframework.http.HttpStatus;
+
+@Getter
+@Builder
+public class ErrorReasonDTO {
+    private final boolean isSuccess;
+    private final HttpStatus httpStatus;
+    private final String code;
+    private final String message;
+}
